requirements.txt
pandas>=1.3.0
fpdf2>=2.7.0
